Core: Refactor ProgressPlugin handling
