+++ lib/cli/src/versions.json
@@ -1,61 +1,61 @@
 {
-  "@storybook/addon-a11y": "6.2.0-alpha.35",
-  "@storybook/addon-actions": "6.2.0-alpha.35",
-  "@storybook/addon-backgrounds": "6.2.0-alpha.35",
-  "@storybook/addon-controls": "6.2.0-alpha.35",
-  "@storybook/addon-cssresources": "6.2.0-alpha.35",
-  "@storybook/addon-design-assets": "6.2.0-alpha.35",
-  "@storybook/addon-docs": "6.2.0-alpha.35",
-  "@storybook/addon-essentials": "6.2.0-alpha.35",
-  "@storybook/addon-events": "6.2.0-alpha.35",
-  "@storybook/addon-google-analytics": "6.2.0-alpha.35",
-  "@storybook/addon-graphql": "6.2.0-alpha.35",
-  "@storybook/addon-jest": "6.2.0-alpha.35",
-  "@storybook/addon-knobs": "6.2.0-alpha.35",
-  "@storybook/addon-links": "6.2.0-alpha.35",
-  "@storybook/addon-queryparams": "6.2.0-alpha.35",
-  "@storybook/addon-storyshots": "6.2.0-alpha.35",
-  "@storybook/addon-storyshots-puppeteer": "6.2.0-alpha.35",
-  "@storybook/addon-storysource": "6.2.0-alpha.35",
-  "@storybook/addon-toolbars": "6.2.0-alpha.35",
-  "@storybook/addon-viewport": "6.2.0-alpha.35",
-  "@storybook/addons": "6.2.0-alpha.35",
-  "@storybook/angular": "6.2.0-alpha.35",
-  "@storybook/api": "6.2.0-alpha.35",
-  "@storybook/aurelia": "6.2.0-alpha.35",
-  "@storybook/builder-webpack4": "6.2.0-alpha.35",
-  "@storybook/builder-webpack5": "6.2.0-alpha.35",
-  "@storybook/channel-postmessage": "6.2.0-alpha.35",
-  "@storybook/channel-websocket": "6.2.0-alpha.35",
-  "@storybook/channels": "6.2.0-alpha.35",
-  "@storybook/cli": "6.2.0-alpha.35",
-  "@storybook/client-api": "6.2.0-alpha.35",
-  "@storybook/client-logger": "6.2.0-alpha.35",
-  "@storybook/codemod": "6.2.0-alpha.35",
-  "@storybook/components": "6.2.0-alpha.35",
-  "@storybook/core": "6.2.0-alpha.35",
-  "@storybook/core-client": "6.2.0-alpha.35",
-  "@storybook/core-common": "6.2.0-alpha.35",
-  "@storybook/core-events": "6.2.0-alpha.35",
-  "@storybook/core-server": "6.2.0-alpha.35",
-  "@storybook/ember": "6.2.0-alpha.35",
-  "@storybook/html": "6.2.0-alpha.35",
-  "@storybook/marionette": "6.2.0-alpha.35",
-  "@storybook/marko": "6.2.0-alpha.35",
-  "@storybook/mithril": "6.2.0-alpha.35",
-  "@storybook/node-logger": "6.2.0-alpha.35",
-  "@storybook/postinstall": "6.2.0-alpha.35",
-  "@storybook/preact": "6.2.0-alpha.35",
-  "@storybook/rax": "6.2.0-alpha.35",
-  "@storybook/react": "6.2.0-alpha.35",
-  "@storybook/riot": "6.2.0-alpha.35",
-  "@storybook/router": "6.2.0-alpha.35",
-  "@storybook/server": "6.2.0-alpha.35",
-  "@storybook/source-loader": "6.2.0-alpha.35",
-  "@storybook/svelte": "6.2.0-alpha.35",
-  "@storybook/theming": "6.2.0-alpha.35",
-  "@storybook/ui": "6.2.0-alpha.35",
-  "@storybook/vue": "6.2.0-alpha.35",
-  "@storybook/vue3": "6.2.0-alpha.35",
-  "@storybook/web-components": "6.2.0-alpha.35"
+  "@storybook/addon-a11y": "6.2.0-beta.0",
+  "@storybook/addon-actions": "6.2.0-beta.0",
+  "@storybook/addon-backgrounds": "6.2.0-beta.0",
+  "@storybook/addon-controls": "6.2.0-beta.0",
+  "@storybook/addon-cssresources": "6.2.0-beta.0",
+  "@storybook/addon-design-assets": "6.2.0-beta.0",
+  "@storybook/addon-docs": "6.2.0-beta.0",
+  "@storybook/addon-essentials": "6.2.0-beta.0",
+  "@storybook/addon-events": "6.2.0-beta.0",
+  "@storybook/addon-google-analytics": "6.2.0-beta.0",
+  "@storybook/addon-graphql": "6.2.0-beta.0",
+  "@storybook/addon-jest": "6.2.0-beta.0",
+  "@storybook/addon-knobs": "6.2.0-beta.0",
+  "@storybook/addon-links": "6.2.0-beta.0",
+  "@storybook/addon-queryparams": "6.2.0-beta.0",
+  "@storybook/addon-storyshots": "6.2.0-beta.0",
+  "@storybook/addon-storyshots-puppeteer": "6.2.0-beta.0",
+  "@storybook/addon-storysource": "6.2.0-beta.0",
+  "@storybook/addon-toolbars": "6.2.0-beta.0",
+  "@storybook/addon-viewport": "6.2.0-beta.0",
+  "@storybook/addons": "6.2.0-beta.0",
+  "@storybook/angular": "6.2.0-beta.0",
+  "@storybook/api": "6.2.0-beta.0",
+  "@storybook/aurelia": "6.2.0-beta.0",
+  "@storybook/builder-webpack4": "6.2.0-beta.0",
+  "@storybook/builder-webpack5": "6.2.0-beta.0",
+  "@storybook/channel-postmessage": "6.2.0-beta.0",
+  "@storybook/channel-websocket": "6.2.0-beta.0",
+  "@storybook/channels": "6.2.0-beta.0",
+  "@storybook/cli": "6.2.0-beta.0",
+  "@storybook/client-api": "6.2.0-beta.0",
+  "@storybook/client-logger": "6.2.0-beta.0",
+  "@storybook/codemod": "6.2.0-beta.0",
+  "@storybook/components": "6.2.0-beta.0",
+  "@storybook/core": "6.2.0-beta.0",
+  "@storybook/core-client": "6.2.0-beta.0",
+  "@storybook/core-common": "6.2.0-beta.0",
+  "@storybook/core-events": "6.2.0-beta.0",
+  "@storybook/core-server": "6.2.0-beta.0",
+  "@storybook/ember": "6.2.0-beta.0",
+  "@storybook/html": "6.2.0-beta.0",
+  "@storybook/marionette": "6.2.0-beta.0",
+  "@storybook/marko": "6.2.0-beta.0",
+  "@storybook/mithril": "6.2.0-beta.0",
+  "@storybook/node-logger": "6.2.0-beta.0",
+  "@storybook/postinstall": "6.2.0-beta.0",
+  "@storybook/preact": "6.2.0-beta.0",
+  "@storybook/rax": "6.2.0-beta.0",
+  "@storybook/react": "6.2.0-beta.0",
+  "@storybook/riot": "6.2.0-beta.0",
+  "@storybook/router": "6.2.0-beta.0",
+  "@storybook/server": "6.2.0-beta.0",
+  "@storybook/source-loader": "6.2.0-beta.0",
+  "@storybook/svelte": "6.2.0-beta.0",
+  "@storybook/theming": "6.2.0-beta.0",
+  "@storybook/ui": "6.2.0-beta.0",
+  "@storybook/vue": "6.2.0-beta.0",
+  "@storybook/vue3": "6.2.0-beta.0",
+  "@storybook/web-components": "6.2.0-beta.0"
 }
