+++ lib/core-common/package.json
@@ -60,6 +60,7 @@
     "@types/glob-base": "^0.3.0",
     "@types/micromatch": "^4.0.1",
     "@types/node": "^14.0.10",
+    "@types/pretty-hrtime": "^1.0.0",
     "babel-loader": "^8.2.2",
     "babel-plugin-macros": "^3.0.1",
     "babel-plugin-polyfill-corejs3": "^0.1.0",
@@ -76,6 +77,7 @@
     "lazy-universal-dotenv": "^3.0.1",
     "micromatch": "^4.0.2",
     "pkg-dir": "^5.0.0",
+    "pretty-hrtime": "^1.0.3",
     "resolve-from": "^5.0.0",
     "ts-dedent": "^2.0.0",
     "util-deprecate": "^1.0.2",
