+++ lib/core-common/src/utils/print-duration.ts
